Move key location and respond to comments
