+++ src/routes/index.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const keys = require('../utils/privateKey.json');
+const keys = require('../keys/privateKey.json');
 const { StatusCodes } = require('http-status-codes');
 
 const db = require('../storage/DataAccess');
@@ -20,5 +20,5 @@ const client = require('../utils/client');
 
 const newClient = new client.Client(keys);
 
-newClient.connectToServer('EHR');
-newClient.connectToServer('KA');
+newClient.connectToServer('http://pathways.mitre.org:8180/fhir');
+newClient.connectToServer('http://pathways.mitre.org:8190/fhir');
