--- src/utils/serverConfig.json
@@ -1,4 +1,0 @@
-{
-	"EHR": "http://pathways.mitre.org:8180/fhir",	
-	"KA": "http://pathways.mitre.org:8190/fhir"
-}
